backend/requirements.txt
# FastAPI
fastapi==0.110.0
uvicorn[standard]==0.27.0
python-multipart==0.0.9

# Database
asyncpg==0.30.0
sqlalchemy[asyncio]==2.0.36
alembic==1.13.1

# Authentication
pyjwt==2.8.0
python-jose[cryptography]==3.3.0
httpx==0.26.0

# Validation
pydantic==2.10.5
pydantic-settings==2.7.1

# Observability
applicationinsights==0.11.10
python-json-logger==2.0.7

# Utils
python-dotenv==1.0.1
